[drivers/pinctrl/sunplus] Updated pinctrl driver of SP7350. 1. Updated pinmux table based on "Q654_Register_v0.3.0.pdf". 2. Changed function/groups name: "CARD0_EMMC" --> "EMMC"

+++ drivers/pinctrl/sunplus/pinctrl_sunplus.h
@@ -102,7 +102,7 @@ struct func_t {
 };
 
 extern struct func_t list_funcs[];
-extern const int list_funcsSZ;
+extern const size_t list_funcsSZ;
 
 extern volatile u32 *moon1_regs;
 #ifdef SUPPORT_PINMUX
